Add RACS2 demo (#26)
Interoperability between cFS and ROS is an important component for space
missions. To increase adoption of Space ROS across different organizations,
we want to add a demo of communication between cFS and ROS to Space ROS.

This commit adds a demonstration of such communication using specifically the
RACS 2 bridge create by JAXA, and the Mars rover demonstration. With this
demo, it is possible to send a command via cFS to ROS that is further sent to
the rover.

This is a very cool demo.